refactor: replace PatientIDScanDialog and SpecimenIDScanDialog with GenericQRScanDialog for improved QR scanning functionality
