+++ package.json
@@ -124,7 +124,6 @@
     "react-intersection-observer": "^9.15.1",
     "react-pdf": "^9.2.1",
     "react-phone-number-input": "^3.4.12",
-    "react-qr-barcode-scanner": "^2.1.2",
     "react-resizable-panels": "^3.0.2",
     "react-webcam": "^7.2.0",
     "react-zoom-pan-pinch": "^3.7.0",
